refactor(web): split the transcript fold and the drafts mirror out of SessionStore
One createStore stays in SessionStore; fold.ts holds the (draft, event)
reducers and Drafts.ts the unsent-message mirror, both mutating the same
draft in the same setState callbacks. Public exports are unchanged.
